Commit Seven

+++ List/views.py
@@ -1,4 +1,4 @@
-from django.shortcuts import render, redirect
+from django.shortcuts import render, redirect, get_object_or_404
 # from django.db.models import F
 from .models import TodoList, ListItem
 from django.contrib import messages
@@ -52,8 +52,8 @@ def is_ajax(request):
 
 def detail(request, pk):
     todo_list = TodoList.objects.get(pk = pk) 
-    list_items = todo_list.list_items.all() # 2nd Option (Reverse Relationship).
-    list_item = todo_list.list_items.filter(item_list = todo_list.pk)
+    # list_items = todo_list.list_items.all() # 2nd Option (Reverse Relationship).
+    list_items = todo_list.list_items.filter(item_list = todo_list.pk)
     # list_items = todo_list.listitem_set.all() # 1st Option (Reverse Relationship). Write the ListItem model in small letter followed by "_set". 
     # data = {}
     # form = ListItemForm()
@@ -65,63 +65,47 @@ def detail(request, pk):
     #         data['item_name'] = form.cleaned_data.get('item_name')
     #         data['status'] = 'ok'
     #         return JsonResponse(data)
+
+    list_item = None
+    if request.method == 'POST':
+        item_form = ListItemForm(request.POST)
+        if item_form.is_valid:
+            list_item = item_form.save(commit=False)
+            list_item.item_list = todo_list
+            list_item.save()
+            return redirect("detail", pk=pk)
+    else:
+        item_form = ListItemForm()
     context = {
         "todo_list": todo_list,
         "list_items": list_items,
         "list_item": list_item,
+        "item_form": item_form
     }
+    return render(request, "detail.html", context)
+
+def delete_list_item(request, list_pk, item_pk):
+    todo_list = TodoList.objects.get(pk = list_pk) 
+    # list_items = todo_list.list_items.filter(item_list = todo_list.pk)
+    list_item = get_object_or_404(ListItem, pk = item_pk)
+
+    # Make sure the item to delete is part of the parent list
+    if list_item.todo_list != todo_list:
+        return redirect("detail", pk = list_pk)
+    
     if request.method == 'POST':
         list_item.delete()
-        return redirect('detail', context)
-    else:
-        return render(request, 'detail.html', context)
+        return redirect('detail', pk = list_pk)
+    
+    return render(request, 'delete-list-item.html', {
+        "list_item": list_item,
+        "todo_list": todo_list
+        })
 
-    # return render(request, 'detail.html', context)
-def new_item(request, pk):
-    todo_list = TodoList.objects.get(pk = pk) 
-    # list_items = todo_list.list_items.all() # 2nd Option (Reverse Relationship).
-    list_items = todo_list.list_items.filter(item_list = todo_list.pk)
-    list_item = None
-    if request.method == 'POST':
-        form = ListItemForm(request.POST)
-        if form.is_valid:
-            list_item = form.save(commit=False)
-            list_item.item_name = todo_list
-            list_item.save()
-            return redirect("new_item")
-    else:
-        form = ListItemForm()
-        return render(request, "new-item.html", {"form": form,})
-
-    # forms.fields['todo_list'].initial = "{{todo_list.pk}}"
-
-def get_items(request, pk):
-    todo_list = TodoList.objects.get(pk = pk) 
-    list_items = todo_list.list_items.all() # 2nd Option (Reverse Relationship).
-    return JsonResponse({"list_items": list(list_items.values())})
-
-# def new_item(request):
-#     if request.method == "POST":
-#         item_name = request.POST["item_name"]
-#         item_list = request.POST["item_list"]
-#         # completed = request.POST["completed"]
-
-#         form = ListItemForm(request.POST)
-#         if form.is_valid:
-#             form.save()
-#         # new_list_item = ListItem(item_name=item_name, item_list=item_list)
-#         # new_list_item.save()
-#         return HttpResponse("New Item Added")
-
-        # '''
-        # todo_list = TodoList.objects.get(pk=pk) 
-        # # list_items = todo_list.listitem_set.all() # 1st Option (Reverse Relationship). Write the ListItem model in small letter followed by "_set". 
-        # list_items = todo_list.list_items.all() # 2nd Option (Reverse Relationship).
-        # context = {
-        #     "todo_list": todo_list,
-        #     "list_items": list_items
-        # }
-        # '''
+# def get_items(request, pk):
+#     todo_list = TodoList.objects.get(pk = pk) 
+#     list_items = todo_list.list_items.all() # 2nd Option (Reverse Relationship).
+#     return JsonResponse({"list_items": list(list_items.values())})
 
 def delete(request, pk):
     todo_list = TodoList.objects.get(pk = pk) 
@@ -129,7 +113,6 @@ def delete(request, pk):
         todo_list.delete()
         return redirect('home')
     else:
-        # messages.info(request, "Are you sure?", {"todo_list": todo_list})
         return render(request, 'delete-list.html', {"todo_list": todo_list})
 
 def deleted(request, pk):
